+++ contracts/adapters/interfaces/IOnboarding.sol
@@ -38,11 +38,11 @@ interface IOnboarding {
 
     function sponsorProposal(
         DaoRegistry dao,
-        uint256 proposalId,
+        uint64 proposalId,
         bytes calldata data
     ) external;
 
-    function cancelProposal(DaoRegistry dao, uint256 proposalId) external;
+    function cancelProposal(DaoRegistry dao, uint64 proposalId) external;
 
-    function processProposal(DaoRegistry dao, uint256 proposalId) external;
+    function processProposal(DaoRegistry dao, uint64 proposalId) external;
 }
